twitter-sentiment-app-v1.1.0/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
PyJWT==2.8.0
pydantic==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.32
alembic==1.13.2
redis==5.0.7
stripe==7.13.0
